Individual child ages: fixing type annoyings

+++ minos/data_generation/generate_stock_pop.py
@@ -275,12 +275,12 @@ def generate_transition_stock(projections, cross_validation):
     #                       var_type='ordinal')
 
     # Set loneliness and ncigs as int
-    data['loneliness'] = data['loneliness'].astype('int64')
-    data['ncigs'] = data['ncigs'].astype('int64')
-    data['neighbourhood_safety'] = data['neighbourhood_safety'].astype('int64')
-    data['nutrition_quality'] = data['nutrition_quality'].astype('int64')
-    data['child_ages'] = data['child_ages'].astype('Int64')
-    data['nkids'] = data['nkids'].astype('Int64')
+    # data['loneliness'] = data['loneliness'].astype('int64')
+    # data['ncigs'] = data['ncigs'].astype('int64')
+    # data['neighbourhood_safety'] = data['neighbourhood_safety'].astype('int64')
+    # data['nutrition_quality'] = data['nutrition_quality'].astype('int64')
+    # data['child_ages'] = data['child_ages'].astype('Int64')
+    # data['nkids'] = data['nkids'].astype('Int64')
     #data['housing_quality'] = data['housing_quality'].astype('int64')
 
     US_utils.save_multiple_files(data, years, "data/final_US/", "")
